// lib/keycloak.d.ts
export type KeycloakOnLoad = 'login-required' | 'check-sso'
export type KeycloakResponseMode = 'query' | 'fragment'
export type KeycloakResponseType = 'code' | 'id_token token' | 'code id_token token'
export type KeycloakFlow = 'standard' | 'implicit' | 'hybrid'
export type KeycloakPkceMethod = 'S256' | false
// ...
export interface KeycloakInitOptions {
  /**
   * Adds a [cryptographic nonce](https://en.wikipedia.org/wiki/Cryptographic_nonce)
   * to verify that the authentication response matches the request.
   * @default true
   */
  useNonce?: boolean

  /**
   *
   * Allow usage of different types of adapters or a custom adapter to make Keycloak work in different environments.
   *
   * The following options are supported:
   * - `default` - Use default APIs that are available in browsers.
   * - `cordova` - Use a WebView in Cordova.
   * - `cordova-native` - Use Cordova native APIs, this is recommended over `cordova`.
   *
   * It's also possible to pass in a custom adapter for the environment you are running Keycloak in. In order to do so extend the `KeycloakAdapter` interface and implement the methods that are defined there.
   *
   * For example:
   *
   * ```ts
   * // Implement the 'KeycloakAdapter' interface so that all required methods are guaranteed to be present.
   * const MyCustomAdapter: KeycloakAdapter = {
   * 	login(options) {
   * 		// Write your own implementation here.
   * 	}
   *
   * 	// The other methods go here...
   * };
   *
   * keycloak.init({
   * 	adapter: MyCustomAdapter,
   * });
   * ```
   */
  adapter?: 'default' | 'cordova' | 'cordova-native' | KeycloakAdapter

  /**
   * Specifies an action to do on load.
   */
  onLoad?: KeycloakOnLoad

  /**
   * Set an initial value for the token.
   */
  token?: string

  /**
   * Set an initial value for the refresh token.
   */
  refreshToken?: string

  /**
   * Set an initial value for the id token (only together with `token` or
   * `refreshToken`).
   */
  idToken?: string

  /**
   * Set an initial value for skew between local time and Keycloak server in
   * seconds (only together with `token` or `refreshToken`).
   */
  timeSkew?: number

  /**
   * Set to enable/disable monitoring login state.
   * @default true
   */
  checkLoginIframe?: boolean

  /**
   * Set the interval to check login state (in seconds).
   * @default 5
   */
  checkLoginIframeInterval?: number

  /**
   * Set the OpenID Connect response mode to send to Keycloak upon login.
   * @default fragment After successful authentication Keycloak will redirect
   *                   to JavaScript application with OpenID Connect parameters
   *                   added in URL fragment. This is generally safer and
   *                   recommended over query.
   */
  responseMode?: KeycloakResponseMode

  /**
   * Specifies a default uri to redirect to after login or logout.
   * This is currently supported for adapter 'cordova-native' and 'default'
   */
  redirectUri?: string

  /**
   * Specifies an uri to redirect to after silent check-sso.
   * Silent check-sso will only happen, when this redirect uri is given and
   * the specified uri is available within the application.
   */
  silentCheckSsoRedirectUri?: string

  /**
   * Specifies whether the silent check-sso should fallback to "non-silent"
   * check-sso when 3rd party cookies are blocked by the browser. Defaults
   * to true.
   */
  silentCheckSsoFallback?: boolean

  /**
   * Set the OpenID Connect flow.
   * @default standard
   */
  flow?: KeycloakFlow

  /**
   * Configures the Proof Key for Code Exchange (PKCE) method to use. This will default to 'S256'.
   * Can be disabled by passing `false`.
   */
  pkceMethod?: KeycloakPkceMethod

  /**
   * Enables logging messages from Keycloak to the console.
   * @default false
   */
  enableLogging?: boolean

  /**
   * Set the default scope parameter to the login endpoint. Use a space-delimited list of scopes.
   * Note that the scope 'openid' will be always be added to the list of scopes by the adapter.
   * Note that the default scope specified here is overwritten if the `login()` options specify scope explicitly.
   */
  scope?: string

  /**
   * Configures how long will Keycloak adapter wait for receiving messages from server in ms. This is used,
   * for example, when waiting for response of 3rd party cookies check.
   *
   * @default 10000
   */
  messageReceiveTimeout?: number

  /**
   * When onLoad is 'login-required', sets the 'ui_locales' query param in compliance with section 3.1.2.1
   * of the OIDC 1.0 specification.
   */
  locale?: string

  /**
   * HTTP method for calling the end_session endpoint. Defaults to 'GET'.
   */
  logoutMethod?: 'GET' | 'POST'
}

export interface KeycloakLoginOptions {
  /**
   * Specifies the scope parameter for the login url
   * The scope 'openid' will be added to the scope if it is missing or undefined.
   */
  scope?: string

  /**
   * Specifies the uri to redirect to after login.
   */
  redirectUri?: string

  /**
   * By default the login screen is displayed if the user is not logged into
   * Keycloak. To only authenticate to the application if the user is already
   * logged in and not display the login page if the user is not logged in, set
   * this option to `'none'`. To always require re-authentication and ignore
   * SSO, set this option to `'login'`. To always prompt the user for consent,
   * set this option to `'consent'`. This ensures that consent is requested,
   * even if it has been given previously.
   */
  prompt?: 'none' | 'login' | 'consent'

  /**
   * If value is `'register'` then user is redirected to registration page,
   * otherwise to login page.
   */
  action?: string

  /**
   * Used just if user is already authenticated. Specifies maximum time since
   * the authentication of user happened. If user is already authenticated for
   * longer time than `'maxAge'`, the SSO is ignored and he will need to
   * authenticate again.
   */
  maxAge?: number

  /**
   * Used to pre-fill the username/email field on the login form.
   */
  loginHint?: string

  /**
   * Sets the `acr` claim of the ID token sent inside the `claims` parameter. See section 5.5.1 of the OIDC 1.0 specification.
   */
  acr?: Acr

  /**
   * Configures the 'acr_values' query param in compliance with section 3.1.2.1
   * of the OIDC 1.0 specification.
   * Used to tell Keycloak what level of authentication the user needs.
   */
  acrValues?: string

  /**
   * Used to tell Keycloak which IDP the user wants to authenticate with.
   */
  idpHint?: string

  /**
   * Sets the 'ui_locales' query param in compliance with section 3.1.2.1
   * of the OIDC 1.0 specification.
   */
  locale?: string

  /**
   * Specifies arguments that are passed to the Cordova in-app-browser (if applicable).
   * Options 'hidden' and 'location' are not affected by these arguments.
   * All available options are defined at https://cordova.apache.org/docs/en/latest/reference/cordova-plugin-inappbrowser/.
   * Example of use: { zoom: "no", hardwareback: "yes" }
   */
  cordovaOptions?: { [optionName: string]: string }
}

export interface KeycloakLogoutOptions {
  /**
   * Specifies the uri to redirect to after logout.
   */
  redirectUri?: string

  /**
   * HTTP method for calling the end_session endpoint. Defaults to 'GET'.
   */
  logoutMethod?: 'GET' | 'POST'
}

export interface KeycloakRegisterOptions extends Omit<KeycloakLoginOptions, 'action'> { }

export interface KeycloakAccountOptions {
  /**
   * Specifies the uri to redirect to when redirecting back to the application.
   */
  redirectUri?: string
}
export interface KeycloakError {
  error: string
  error_description: string
}

export interface KeycloakRedirectUriOptions {
  /**
   * Specifies the uri to redirect to after login.
   */
  redirectUri?: string
}

export interface KeycloakAdapter {
  login: (options?: KeycloakLoginOptions) => Promise<void>
  logout: (options?: KeycloakLogoutOptions) => Promise<void>
  register: (options?: KeycloakRegisterOptions) => Promise<void>
  accountManagement: () => Promise<void>
  redirectUri: (options?: KeycloakRedirectUriOptions) => string
}